[9] Remove BetterCodeHub
This repository does not get any value from the BetterCodeHub rating.

These guidelines are covered by CodeFactor and CodeCov in more detail.
* Write code once
* Automate Tests
* Write Clean Code

These guidelines are subjective and can easily be moderated in code reviews. Putting arbitrary hard limits on these does not really do anything for code maintainability.
* Write Short Units of Code
* Write Simple Units of Code
* Keep Unit Interfaces Small
* Separate Concerns in Modules
* Couple Architecture Components Loosely

These guidelines will always be at odds with what this repository is: A collection of individual libraries that are not necessarily related.
Perhaps this is an indication that they should be split into different repositories, but that is a different point altogether.
* Keep Architecture Components Balanced
* Keep your codebase small

https://github.com/Dangerdan9631/DangrLib/issues/9